Move types to PanView
