+++ src/incubator/Dialog/index.tsx
@@ -2,8 +2,7 @@ import React, {PropsWithChildren, useMemo, useCallback} from 'react';
 import {StyleSheet, StyleProp, ViewStyle} from 'react-native';
 import {useSharedValue, withTiming, useAnimatedStyle} from 'react-native-reanimated';
 import TransitionView, {TransitionViewAnimationType} from '../TransitionView';
-import PanView from '../panView';
-import {PanningDirections, PanningDirectionsEnum} from '../panView/panningUtil';
+import PanView, {PanningDirections, PanningDirectionsEnum} from '../panView';
 import View from '../../components/view';
 import Modal, {ModalProps} from '../../components/modal';
 import {AlignmentModifiers} from '../../commons/modifiers';
